Move message classes into the `type` package
