+++ client/src/main/java/org/spine3/type/CommandClass.java
@@ -18,10 +18,12 @@
  * OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
 
-package org.spine3.base;
+package org.spine3.type;
 
 import com.google.common.collect.ImmutableSet;
 import com.google.protobuf.Message;
+import org.spine3.base.Command;
+import org.spine3.base.Commands;
 
 import java.util.Arrays;
