+++ server/src/main/java/org/spine3/server/command/CommandDispatcher.java
@@ -20,9 +20,9 @@
 
 package org.spine3.server.command;
 
-import org.spine3.base.CommandClass;
 import org.spine3.envelope.CommandEnvelope;
 import org.spine3.server.bus.MessageDispatcher;
+import org.spine3.type.CommandClass;
 
 /**
  * Delivers commands to their handlers.
