+++ server/src/main/java/org/spine3/server/command/error/CommandExpiredException.java
@@ -21,11 +21,11 @@ package org.spine3.server.command.error;
 
 import com.google.protobuf.Message;
 import org.spine3.base.Command;
-import org.spine3.base.CommandClass;
 import org.spine3.base.CommandValidationError;
 import org.spine3.base.Commands;
 import org.spine3.base.Error;
 import org.spine3.protobuf.TypeName;
+import org.spine3.type.CommandClass;
 
 import static java.lang.String.format;
