+++ server/src/main/java/org/spine3/server/command/error/InvalidCommandException.java
@@ -22,13 +22,13 @@ package org.spine3.server.command.error;
 
 import com.google.protobuf.Message;
 import org.spine3.base.Command;
-import org.spine3.base.CommandClass;
 import org.spine3.base.CommandContext;
 import org.spine3.base.CommandValidationError;
 import org.spine3.base.Commands;
 import org.spine3.base.Error;
 import org.spine3.base.ValidationError;
 import org.spine3.protobuf.TypeName;
+import org.spine3.type.CommandClass;
 import org.spine3.validate.ConstraintViolation;
 
 import static java.lang.String.format;
