+++ server/src/main/java/org/spine3/server/event/error/UnsupportedEventException.java
@@ -21,9 +21,9 @@ package org.spine3.server.event.error;
 
 import com.google.protobuf.Message;
 import org.spine3.base.Error;
-import org.spine3.base.EventClass;
 import org.spine3.base.EventValidationError;
 import org.spine3.protobuf.TypeName;
+import org.spine3.type.EventClass;
 
 /**
  * Exception that is thrown when unsupported event is obtained
