+++ server/src/main/java/org/spine3/server/procman/ProcessManagerRepository.java
@@ -23,10 +23,8 @@ package org.spine3.server.procman;
 import com.google.protobuf.Message;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
-import org.spine3.base.CommandClass;
 import org.spine3.base.CommandContext;
 import org.spine3.base.Event;
-import org.spine3.base.EventClass;
 import org.spine3.base.EventContext;
 import org.spine3.base.Events;
 import org.spine3.envelope.CommandEnvelope;
@@ -36,6 +34,8 @@ import org.spine3.server.command.CommandDispatcher;
 import org.spine3.server.entity.EventDispatchingRepository;
 import org.spine3.server.entity.idfunc.GetTargetIdFromCommand;
 import org.spine3.server.event.EventBus;
+import org.spine3.type.CommandClass;
+import org.spine3.type.EventClass;
 
 import javax.annotation.CheckReturnValue;
 import javax.annotation.Nullable;
