+++ server/src/main/java/org/spine3/server/projection/ProjectionRepository.java
@@ -30,7 +30,6 @@ import io.grpc.stub.StreamObserver;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.spine3.base.Event;
-import org.spine3.base.EventClass;
 import org.spine3.base.EventContext;
 import org.spine3.protobuf.TypeName;
 import org.spine3.server.BoundedContext;
@@ -43,6 +42,7 @@ import org.spine3.server.stand.StandFunnel;
 import org.spine3.server.storage.RecordStorage;
 import org.spine3.server.storage.Storage;
 import org.spine3.server.storage.StorageFactory;
+import org.spine3.type.EventClass;
 
 import javax.annotation.Nonnull;
 import javax.annotation.Nullable;
