+++ server/src/test/java/org/spine3/server/command/CommandBusShould.java
@@ -24,7 +24,6 @@ import com.google.common.base.Optional;
 import io.grpc.stub.StreamObserver;
 import org.junit.Test;
 import org.spine3.base.Command;
-import org.spine3.base.CommandClass;
 import org.spine3.base.CommandContext;
 import org.spine3.base.CommandValidationError;
 import org.spine3.base.Error;
@@ -38,6 +37,7 @@ import org.spine3.server.users.CurrentTenant;
 import org.spine3.test.Tests;
 import org.spine3.test.command.AddTask;
 import org.spine3.test.command.CreateProject;
+import org.spine3.type.CommandClass;
 import org.spine3.users.TenantId;
 
 import java.util.Set;
