+++ server/src/test/java/org/spine3/server/procman/AbstractCommandRouterShould.java
@@ -28,7 +28,6 @@ import com.google.protobuf.Message;
 import com.google.protobuf.StringValue;
 import org.junit.Before;
 import org.spine3.base.Command;
-import org.spine3.base.CommandClass;
 import org.spine3.base.CommandContext;
 import org.spine3.base.Commands;
 import org.spine3.envelope.CommandEnvelope;
@@ -37,6 +36,7 @@ import org.spine3.server.BoundedContext;
 import org.spine3.server.command.CommandBus;
 import org.spine3.server.command.CommandDispatcher;
 import org.spine3.test.TestCommandFactory;
+import org.spine3.type.CommandClass;
 
 import javax.annotation.Nullable;
 import java.util.List;
