+++ server/src/test/java/org/spine3/server/procman/ProcessManagerRepositoryShould.java
@@ -30,10 +30,8 @@ import org.junit.After;
 import org.junit.Before;
 import org.junit.Test;
 import org.mockito.ArgumentCaptor;
-import org.spine3.base.CommandClass;
 import org.spine3.base.CommandContext;
 import org.spine3.base.Event;
-import org.spine3.base.EventClass;
 import org.spine3.base.EventContext;
 import org.spine3.base.Events;
 import org.spine3.envelope.CommandEnvelope;
@@ -58,6 +56,8 @@ import org.spine3.test.procman.event.TaskAdded;
 import org.spine3.testdata.Sample;
 import org.spine3.testdata.TestBoundedContextFactory;
 import org.spine3.testdata.TestEventBusFactory;
+import org.spine3.type.CommandClass;
+import org.spine3.type.EventClass;
 
 import java.lang.reflect.InvocationTargetException;
 import java.util.List;
