+++ server/src/test/java/org/spine3/server/procman/ProcessManagerShould.java
@@ -28,10 +28,8 @@ import io.grpc.stub.StreamObserver;
 import org.junit.Before;
 import org.junit.Test;
 import org.spine3.base.Command;
-import org.spine3.base.CommandClass;
 import org.spine3.base.CommandContext;
 import org.spine3.base.Event;
-import org.spine3.base.EventClass;
 import org.spine3.base.EventContext;
 import org.spine3.base.Events;
 import org.spine3.envelope.CommandEnvelope;
@@ -52,6 +50,8 @@ import org.spine3.test.procman.event.ProjectCreated;
 import org.spine3.test.procman.event.ProjectStarted;
 import org.spine3.test.procman.event.TaskAdded;
 import org.spine3.testdata.Sample;
+import org.spine3.type.CommandClass;
+import org.spine3.type.EventClass;
 
 import java.lang.reflect.InvocationTargetException;
 import java.util.List;
