+++ server/src/test/java/org/spine3/server/projection/ProjectionRepositoryShould.java
@@ -29,7 +29,6 @@ import com.google.protobuf.StringValue;
 import org.junit.Before;
 import org.junit.Test;
 import org.spine3.base.Event;
-import org.spine3.base.EventClass;
 import org.spine3.base.EventContext;
 import org.spine3.base.EventId;
 import org.spine3.base.Events;
@@ -55,6 +54,7 @@ import org.spine3.test.projection.event.TaskAdded;
 import org.spine3.testdata.Sample;
 import org.spine3.testdata.TestBoundedContextFactory;
 import org.spine3.testdata.TestEventBusFactory;
+import org.spine3.type.EventClass;
 
 import java.util.Collection;
 import java.util.LinkedList;
